requirements.txt
llama-cpp-python>=0.2.0
tqdm>=4.66.0
